Workaround to use an upstream fix the zombie problem
The zombie processes problem is a currently open issue of
opencontainers:

https://github.com/opencontainers/runc/issues/1443

Currently there is a PR that I've tested to work with our use cases:

https://github.com/opencontainers/runc/pull/1506

But Glide doesn't seem to be able to get a version from a PR; and if I add
LittleLightLittleFire's to the dependencies (and change the imports in the Go
files) glide then fails with some error related to the other parts of
libopencontainers and the fork having the same dependencies.

So, this (ugly but simple) workaround in the Makefile was created to use the
upstream PR so the machine-learning department can continue slurping all the
code in the world while the upstream PR is merged.