daemon: fix the context passed to the driver factory; fixes #253
Previous changes added a context argument to newDriverPool, which is
called when the first request for the driver is received. By mistake,
the context of this first request is captured by the closure and is used
to create all other driver instances. Thus, if the first request
specifies the context timeout and fails for some reason, all future
requests will fails with "context cancelled" error.

Instead, we need to pass the real context from the user request down to
the pool's driver factory function.

This is not an ideal way of managing drivers - pool should be managed by
a separate goroutine. This will be implemented later. This is merely a
hotfix for the real issue.

Signed-off-by: Denys Smirnov <denys@sourced.tech>