daemon: simplify the code for moving average scaling policy
The code for moving average scaling policy is a bit more complex than it
can be. It defines and checks the state variables to know when the
internal ring buffer becomes full and "wraps" around.

Instead of adding those state variables, the capacity of the slice can
be used to track the desired buffer size, and the length can track the
number of elements in the buffer.

The change removes unnecessary state variables and uses the buffer
capacity, as described above. The code becomes cleaner, since the length
of the buffer can be used unconditionally, and the buffer can be
iterated with the range statement.

Signed-off-by: Denys Smirnov <denys@sourced.tech>