daemon: switch from sync.Map to a regular map with a mutex
Daemon makes heavy use of lock-free operations: all the counters are
atomic, and the drivers map is lock-free as well (sync.Map). This was
probably done to avoid lock contention.

But in fact, the contention will still exist because each operation on
the sync.Map is accompanied with an operation on the driver queue, which
uses Go channel under the hood. Channels are definitely slower than
mutexes, and much slower than atomic operations. Thus, the code cannot
really use the speed and contention avoidance offered by sync.Map.

This change replaces sync.Map with a regular Go map with a RW mutex to
simplify the code.

Signed-off-by: Denys Smirnov <denys@sourced.tech>