daemon: use context with opentracing enabled when creating the pool
The context passed to newDriverPool creates the new OpenTracing span.
Previously, the context was not passed further, thus there was no need
to use the context with a new span attached. Now the context is passed
to the driver factory function, thus we should use the wrapped context.

Also add a new tracing span to the factory function as well.

Signed-off-by: Denys Smirnov <denys@sourced.tech>