daemon: rewrite the driver connection pool
An old pool had some problems like leaking goroutines, leaving dead
drivers to hang for too long and unpredictable scaling behavior.

The new driver pool introduced in this change should be safer an more
predictable. It may also fix issues with the scaling that we discovered.

New implementation still runs a scaling goroutine that sets a target
number of instances. But instead of directly applying this number by
killing or starting instances, it delegates instance management to a
separate goroutine.

This second goroutine also accepts requests for idle drivers. It listens
on multiple event channels and it should be now clean from the code what
decision it would make on each event.

Also, an old implementation was using a client's request context when
starting new instances. In this implementation a third goroutine is
responsible for starting instances in the background.

Signed-off-by: Denys Smirnov <denys@sourced.tech>